Initial Commit
The result of 3 hours, Will do more. Im working towards making an env
pulsr and other applications we do can scale and thrive in a "shared"
ecosystem. e.g adding one server will benefit the whole cluster and all
the applications running on it. This is just for HTTP traffic.